POLIO-1937: regroup lqas im types in one file
