+++ plugins/polio/js/src/domains/LQAS-IM/IM/CountryOverview/ImCountryListOverview.tsx
@@ -4,8 +4,7 @@ import React, {
     useMemo,
     useState,
 } from 'react';
-import { useSafeIntl } from 'bluesquare-components';
-import { cloneDeep, sortBy } from 'lodash';
+import MapIcon from '@mui/icons-material/Map';
 import {
     TableContainer,
     Table as MuiTable,
@@ -18,7 +17,9 @@ import {
     Typography,
 } from '@mui/material';
 import { makeStyles } from '@mui/styles';
-import MapIcon from '@mui/icons-material/Map';
+import { useSafeIntl } from 'bluesquare-components';
+import { cloneDeep, sortBy } from 'lodash';
+import MESSAGES from '../../../../constants/messages';
 import {
     OK_COLOR,
     FAIL_COLOR,
@@ -26,13 +27,12 @@ import {
 } from '../../../../styles/constants';
 
 import { useStyles } from '../../../../styles/theme';
-import MESSAGES from '../../../../constants/messages';
 import { TablePlaceHolder } from '../../../Campaigns/Scope/Scopes/TablePlaceHolder';
 import { TableText } from '../../../Campaigns/Scope/Scopes/TableText';
 import { IN_SCOPE } from '../../shared/constants';
-import { IM_FAIL, IM_PASS, IM_WARNING } from '../constants';
 import { HasLocationIcon } from '../../shared/HasLocationIcon';
 import { findRegionShape } from '../../shared/utils';
+import { IM_FAIL, IM_PASS, IM_WARNING } from '../constants';
 
 type SortFocus =
     | 'DISTRICT'
