+++ plugins/polio/js/src/domains/LQAS-IM/LQAS/CountryOverview/LqasOverviewContainer.tsx
@@ -2,19 +2,16 @@ import React, { FunctionComponent, useMemo } from 'react';
 
 import { Divider, Paper } from '@mui/material';
 import { DropdownOptions } from '../../../../../../../../hat/assets/js/apps/Iaso/types/utils';
-import {
-    Campaign,
-    ConvertedLqasImData,
-    Side,
-} from '../../../../constants/types';
+import { Campaign, MapShapes, Side } from '../../../../constants/types';
 import { baseUrls } from '../../../../constants/urls';
-import { getLqasImMapLayer, LqasImMapLayer } from '../../IM/utils';
+import { getLqasImMapLayer } from '../../IM/utils';
 import { LIST, LqasIMView, MAP } from '../../shared/constants';
-import { MapShapes, useMapShapes } from '../../shared/hooks/api/useMapShapes';
+import { useMapShapes } from '../../shared/hooks/api/useMapShapes';
 import { useLqasImMapHeaderData } from '../../shared/hooks/useLqasImMapHeaderData';
 import { LqasImMapHeader } from '../../shared/Map/LqasImMapHeader';
 import { LqasImTabs } from '../../shared/Tabs/LqasImTabs';
 import { useLqasImTabState } from '../../shared/Tabs/useLqasImTabState';
+import { ConvertedLqasImData, LqasImMapLayer } from '../../types';
 import { LqasCountryListOverview } from './LqasCountryListOverview';
 import { LqasCountryMap } from './LqasCountryMap';
 import { LqasSummary } from './LqasSummary';
