+++ plugins/polio/js/src/domains/LQAS-IM/LQAS/Graphs/CaregiversTableHeader.tsx
@@ -4,13 +4,13 @@ import { Box, Typography, Paper } from '@mui/material';
 import { makeStyles } from '@mui/styles';
 import { useSafeIntl } from 'bluesquare-components';
 import MESSAGES from '../../../../constants/messages';
-import { ConvertedLqasImData } from '../../../../constants/types';
 import {
     accessArrayRound,
     convertStatToPercent,
     totalCaregivers,
     totalCaregiversInformed,
 } from '../../shared/LqasIm';
+import { ConvertedLqasImData } from '../../types';
 
 type Props = {
     campaign?: string;
