+++ plugins/polio/js/src/domains/LQAS-IM/shared/HasLocationIcon.tsx
@@ -1,9 +1,8 @@
 import React, { FunctionComponent } from 'react';
-import { defineMessages } from 'react-intl';
-import { Tooltip, Box } from '@mui/material';
 import LocationDisabledIcon from '@mui/icons-material/LocationDisabled';
-
+import { Tooltip, Box } from '@mui/material';
 import { useSafeIntl } from 'bluesquare-components';
+import { defineMessages } from 'react-intl';
 import ShapeSvg from '../../../../../../../hat/assets/js/apps/Iaso/components/svg/ShapeSvgComponent';
 import { Shape } from '../../../constants/types';
