+++ plugins/polio/js/src/domains/LQAS-IM/shared/Map/LqasImCountryMap.tsx
@@ -2,12 +2,11 @@ import React, { FunctionComponent } from 'react';
 import { Box } from '@mui/material';
 import { LoadingSpinner } from 'bluesquare-components';
 import { GreyHashedPattern } from '../../../../../../../../hat/assets/js/apps/Iaso/components/maps/HashedPatterns/GreyHashedPattern';
-import { ConvertedLqasImData } from '../../../../constants/types';
 import { LegendItem, defaultShapeStyle } from '../../../../utils/index';
 import { MapComponent } from '../../../Campaigns/MapComponent/MapComponent';
 import { MapLegend } from '../../../Campaigns/MapComponent/MapLegend';
 import { MapLegendContainer } from '../../../Campaigns/MapComponent/MapLegendContainer';
-import { LqasImMapLayer } from '../../IM/utils';
+import { ConvertedLqasImData, LqasImMapLayer } from '../../types';
 import { HASHED_MAP_PATTERN } from '../constants';
 import { makePopup } from './LqasImPopUp';
 import { ScopeAndDNFDisclaimer } from './ScopeAndDNFDisclaimer';
