+++ plugins/polio/js/src/domains/LQAS-IM/shared/Map/LqasImDates.tsx
@@ -1,11 +1,11 @@
+import React, { FunctionComponent } from 'react';
+import InfoOutlinedIcon from '@mui/icons-material/InfoOutlined';
 import { Grid, Tooltip } from '@mui/material';
 import { makeStyles } from '@mui/styles';
-import React, { FunctionComponent } from 'react';
 import { useSafeIntl } from 'bluesquare-components';
-import InfoOutlinedIcon from '@mui/icons-material/InfoOutlined';
 import MESSAGES from '../../../../constants/messages';
 import { OK_COLOR, FAIL_COLOR } from '../../../../styles/constants';
-import { LqasImRefDate } from './LqasImMapHeader';
+import { LqasImRefDate } from '../../types';
 
 type Props = {
     type: 'start' | 'end';
