+++ plugins/polio/js/src/domains/LQAS-IM/shared/Map/ScopeAndDNFDisclaimer.tsx
@@ -1,13 +1,14 @@
+import React, { FunctionComponent } from 'react';
 import { Box, Grid, Typography } from '@mui/material';
 import { makeStyles } from '@mui/styles';
-import React, { FunctionComponent } from 'react';
 import { useSafeIntl } from 'bluesquare-components';
 import MESSAGES from '../../../../constants/messages';
 import { findScopeIds } from '../../../../utils';
+import { LqasImDebugData } from '../../types';
 
 type Props = {
     campaign?: string;
-    data: Record<string, { hasScope: boolean; districtsNotFound: string[] }>;
+    data: LqasImDebugData;
     campaigns?: any;
     round?: number;
 };
