+++ plugins/polio/js/src/domains/LQAS-IM/shared/charts/PercentageChartWithTitle.tsx
@@ -1,13 +1,13 @@
 import React, { FunctionComponent } from 'react';
 import { Box, Typography } from '@mui/material';
 import { LoadingSpinner } from 'bluesquare-components';
-import { FormatLqasDataForChartResult } from '../../LQAS/utils';
+import { LqasDataForChart } from '../../types';
 import { PercentageBarChart } from './PercentageBarChart';
 import { GraphTooltipFormatter } from './PercentageBarChart/types';
 
 type Props = {
     title?: string;
-    data: FormatLqasDataForChartResult[];
+    data: LqasDataForChart[];
     tooltipFormatter: GraphTooltipFormatter;
     isLoading: boolean;
     chartKey: string;
