+++ plugins/polio/js/src/domains/LQAS-IM/shared/constants.ts
@@ -1,5 +1,5 @@
-import { LqasIMtype } from '../../../constants/types';
 import { NO_DATA_COLOR, NO_DATA_COLOR_2 } from '../../../styles/constants';
+import { LqasIMType } from '../types';
 
 export const paperElevation = 2;
 export const HASHED_MAP_PATTERN = 'greyStripes';
@@ -10,8 +10,8 @@ export const COUNTRY = 'country';
 export const LIST = 'list';
 export const MAP = 'map';
 export const LqasIMView = {
-    lqas: 'lqas' as LqasIMtype,
-    imIHH: 'imIHH' as LqasIMtype,
-    imOHH: 'imOHH' as LqasIMtype,
-    imGlobal: 'imGlobal' as LqasIMtype,
+    lqas: 'lqas' as LqasIMType,
+    imIHH: 'imIHH' as LqasIMType,
+    imOHH: 'imOHH' as LqasIMType,
+    imGlobal: 'imGlobal' as LqasIMType,
 };
