+++ plugins/polio/js/src/domains/LQAS-IM/shared/hooks/useDebugData.ts
@@ -1,10 +1,10 @@
 import { useMemo } from 'react';
-import { LqasImData } from '../../../../constants/types';
+import { LqasImData, LqasImDebugData } from '../../types';
 
 export const makeDebugData = (
     data: LqasImData | undefined,
     campaign?: string,
-): Record<string, { hasScope: boolean; districtsNotFound: string[] }> => {
+): LqasImDebugData => {
     if (!data?.stats || !campaign || !data.stats[campaign]) return {};
     return {
         [campaign]: {
