--- plugins/polio/js/src/domains/LQAS-IM/shared/types/types.ts
@@ -1,1 +1,0 @@
-export type LQASIMType = 'lqas' | 'imIHH' | 'imOHH' | 'imGlobal';
